refactor(provisioner): split NGFW DB helpers into provisioner_db_ngfw (#878)
Clear the S104 (>500 lines) smell on provisioner_db.py (501 -> 343 lines) by
extracting the cohesive NGFW range-attachment read/write helpers
(get_user_ngfw_data, get_ngfw_data_by_request_id, and the attachment-record
helpers) into a focused provisioner_db_ngfw.py. They keep reaching the psycopg
factory and instance-state writer through `main`, so the `main.<name>` call
sites and `patch("main.X")` test seams are unchanged; main re-exports the new
module's surface. Full provisioner suite: 753 passed, 8 skipped.
